Move logic from service to factories.
